Generate timestamp should have been long instead of int

+++ Assets/Scripts/Tools/Utils.cs
@@ -105,11 +105,11 @@ public static class Utils
     {
         return finalPrice / (1 + (percentage / 100));
     }
-    public static int CurrentTimestamp()
+    public static long CurrentTimestamp() // TODO: investigate why it return negative value, probably has something to do with int becuase it should have been long
     {
         DateTime epochStart = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
         TimeSpan elapsedTime = DateTime.UtcNow - epochStart;
-        return (int)elapsedTime.TotalMilliseconds;
+        return (long)elapsedTime.TotalMilliseconds;
     }
     /* ModifiedSigmoid
      * Sigmoid is a generic function that takes any numbers from -infinity to infinity and squash it down to value between 0 to 1
